Made some clean code changes
